requirements.txt
face_recognition==1.3.0
opencv-python==4.7.0.72
numpy==1.24.3
dlib==19.24.0
pillow==9.5.0
